購入機能

+++ app/controllers/tweets_controller.rb
@@ -1,5 +1,7 @@
 class TweetsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show, :search]
+  before_action :set_item, only: [:show, :edit, :update]
+  before_action :move_to_index, only: [:edit, :update]
 
   def index
     @tweets = Tweet.includes(:user).order("created_at DESC")
@@ -19,20 +21,19 @@ class TweetsController < ApplicationController
   end
 
   def show
-    @tweet = Tweet.find(params[:id])
     @comment = Comment.new
     @comments = @tweet.comments.includes(:user)
   end
 
   def edit
-    @tweet = Tweet.find(params[:id])
+    move_to_root_path
     unless current_user.id == @tweet.user_id
       redirect_to action: :index
     end
   end
 
   def update
-    @tweet = Tweet.find(params[:id])
+    move_to_root_path
     if @tweet.update(tweet_params)
       redirect_to tweet_path(@tweet.id)
     else
@@ -42,6 +43,7 @@ class TweetsController < ApplicationController
 
   def destroy
     tweet = Tweet.find(params[:id])
+    redirect_to root_path if tweet.record.present?
     tweet.destroy
     redirect_to root_path
   end
@@ -52,6 +54,18 @@ class TweetsController < ApplicationController
 
   private
   def tweet_params
-    params.require(:tweet).permit(:store_name, :store_type, :information, :place_area_id, :phone, :image).merge(user_id: current_user.id)
+    params.require(:tweet).permit(:store_name, :store_type, :information, :place_area_id, :phone, :image, :price).merge(user_id: current_user.id)
+  end
+
+  def set_item
+    @tweet = Tweet.find(params[:id])
+  end
+
+  def move_to_index
+    redirect_to action: :index unless current_user.id == @tweet.user_id
+  end
+
+  def move_to_root_path
+    redirect_to root_path if @tweet.record.present?
   end
 end
